+++ index.js
@@ -1,82 +1,15 @@
 'use strict';
 
-var valid = {};
-
-// ----- Regexp checks from is_js
-//		--	Modified url to include port: https://gist.github.com/dperini/729294
-//		--	Modified affirmative to include 'on' and allowed any capitalization
-//		--	Added negatory
-//		--	Improved creditCard regexp allows for more card types
-//				- http://stackoverflow.com/a/23231321
-// ---------------------------------------
-valid._regexps = {
-  affirmative: /^(?:1|t(?:rue)?|y(?:es)?|on|ok(?:ay)?)$/,
-  alphabetic: /^[A-Za-z]+$/,
-  alphaNumeric: /^[A-Za-z0-9]+$/,
-  creditCard: {
-  	visa: /^4[0-9]{12}(?:[0-9]{3})?$/,
-  	mastercard: /^5[1-5][0-9]{14}$/,
-  	amex: /^3[47][0-9]{13}$/,
-  	carteBlanche: /^389[0-9]{11}$|^380[0-9]{11}$/,
-  	dinersClub: /^3(?:0[0-5]|[68][0-9])[0-9]{11}$/,
-  	discover: /^65[4-9][0-9]{13}|64[4-9][0-9]{13}|6011[0-9]{12}|(622(?:12[6-9]|1[3-9][0-9]|[2-8][0-9][0-9]|9[01][0-9]|92[0-5])[0-9]{10})$/,
-  	jcb: /^(?:2131|1800|35\d{3})\d{11}$/,
-  	lasercard: /^(6304|6706|6709|6771)[0-9]{12,15}$/,
-  	maestro: /^(5018|5020|5038|6304|6759|6761|6763|6799)[0-9]{8,19}$/,
-  	solo: /^(6334|6767)[0-9]{12}|(6334|6767)[0-9]{14}|(6334|6767)[0-9]{15}$|^(4903|4905|4911|4936|6333|6759)[0-9]{12}|(4903|4905|4911|4936|6333|6759)[0-9]{14}|(4903|4905|4911|4936|6333|6759)[0-9]{15}|564182[0-9]{10}|564182[0-9]{12}|564182[0-9]{13}|633110[0-9]{10}|633110[0-9]{12}|633110[0-9]{13}$/,
-  	unionpay: /^(62[0-9]{14,17})$/
-  },
-  email: /^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$/i,
-  negatory: /^(?:1|f(?:alse)?|n(?:o)?|off)$/,
-  url: /^(?:(?:https?|ftp):\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/\S*)?$/i,
-  zip: /^\d{5}(-\d{4})?$/
-};
-
-
-// --------------------------------- string methods ---------------------------------
-
-// ----- private typechecker method
-// ---------------------------------------
-valid._isString = function(str) {
-
-	if (typeof str === 'string' || str instanceof String) {
-		return true;
-	}
-
-	return false;
-
-};
-
-// ----- private regexp test methods
-// ---------------------------------------
-valid._testRegexp = function(regexp, str) {
-
-	if (!this._isString(str)) {
-		return false;
-	}
-
-	str = str.toLowerCase();
-
-	return this._regexps[regexp].test(str);
-
-};
-
-valid._testCardRegexp = function(regexp, str) {
-	
-	if (!this._isString(str)) {
-		return false;
-	}
-
-	return this._regexps.creditCard[regexp].test(str);
-
-};
+var valid = module.exports = {};
 
 
 // ----- affirmative
 // ---------------------------------------
 valid.affirmative = function(str) {
 
-	return this._testRegexp('affirmative', str);
+	var regex = /^(?:1|t(?:rue)?|y(?:es)?|on|ok(?:ay)?)$/;
+
+	return regex.test(str.toLowerCase());
 
 };
 
@@ -85,7 +18,9 @@ valid.affirmative = function(str) {
 // ---------------------------------------
 valid.alphaNumeric = function(str) {
 
-	return this._testRegexp('alphaNumeric', str);
+	var regex = /^[A-Za-z0-9]+$/;
+
+	return regex.test(str.toLowerCase());
 
 };
 
@@ -94,7 +29,9 @@ valid.alphaNumeric = function(str) {
 // ---------------------------------------
 valid.alpha = function(str) {
 
-	return this._testRegexp('alphabetic', str);
+	var regex = /^[A-Za-z]+$/;
+
+	return regex.test(str.toLowerCase());
 
 };
 
@@ -103,7 +40,9 @@ valid.alpha = function(str) {
 // ---------------------------------------
 valid.email = function(str) {
 
-	return this._testRegexp('email', str);
+	var regex = /^((([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+(\.([a-z]|\d|[!#\$%&'\*\+\-\/=\?\^_`{\|}~]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])+)*)|((\x22)((((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(([\x01-\x08\x0b\x0c\x0e-\x1f\x7f]|\x21|[\x23-\x5b]|[\x5d-\x7e]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(\\([\x01-\x09\x0b\x0c\x0d-\x7f]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))))*(((\x20|\x09)*(\x0d\x0a))?(\x20|\x09)+)?(\x22)))@((([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|\d|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))\.)+(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])|(([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])([a-z]|\d|-|\.|_|~|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])*([a-z]|[\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])))$/i;
+
+	return regex.test(str.toLowerCase());
 
 };
 
@@ -112,7 +51,10 @@ valid.email = function(str) {
 // ---------------------------------------
 valid.negatory = function(str) {
 
-	return this._testRegexp('negatory', str);
+	var regex = /^(?:1|f(?:alse)?|n(?:o)?|off)$/;
+
+	return regex.test(str.toLowerCase());
+
 
 };
 
@@ -121,9 +63,7 @@ valid.negatory = function(str) {
 // ---------------------------------------
 valid.numeric = function(str) {
 
-	if (!this._isString(str)) {
-		return false;
-	}
+	str = str.toLowerCase();
 
 	if (str === '') {
 		return false;
@@ -138,21 +78,15 @@ valid.numeric = function(str) {
 // ---------------------------------------
 valid.value = function(str) {
 
-	if (this._isString(str)) {
+	str = str.toLowerCase().replace(',', '');
 
-		str = str.replace(',', '');
-		var numbers = ['0','1','2','3','4','5','6','7','8','9', '-'];
-		
-		// for $100 and #123, etc.
-		if (numbers.indexOf(str[0]) === -1) {
-			str = str.slice(1);
-		}
-
+	var numbers = ['0','1','2','3','4','5','6','7','8','9', '-'];
+	
+	// for $100 and #123, etc.
+	if (numbers.indexOf(str[0]) === -1) {
+		str = str.slice(1);
 	}
 
-	else {
-		return false;
-	}
 
 	return !isNaN(parseInt(str), 10);
 
@@ -163,22 +97,34 @@ valid.value = function(str) {
 // ---------------------------------------
 valid.url = function(str) {
 
-	return this._testRegexp('url', str);
+	var regex = /^(?:(?:https?|ftp):\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/\S*)?$/i;
+
+	return regex.test(str.toLowerCase());
 
 };
 
 
 // ----- zip code
+//		--	1. input must be a string
+//		--	2. check number
+//		--	3. perform regex test
 // ---------------------------------------
 valid.zipCode = function(str) {
 
+	// [1] throws here if not a string
+	str = str.toLowerCase();
+
+	// [2] every zip code begins with a number, and none are < 00501
 	var numeric = parseInt(str, 10);
 
 	if (isNaN(numeric) || numeric < 501) {
 		return false;
 	}
 
-	return this._testRegexp('zip', str);
+	// [3]
+	var regex = /^\d{5}(-\d{4})?$/;
+
+	return regex.test(str);
 
 };
 
@@ -187,15 +133,11 @@ valid.zipCode = function(str) {
 
 // ----- credit card check constructor
 // ---------------------------------------
-function CreditCard(name) {
+function CreditCard(regex) {
 
 	return function(str) {
 
-		if (!valid._testCardRegexp(name, str)) {
-			return false;
-		}
-
-		return true;
+		return regex.test(str.toLowerCase());
 
 	};
 
@@ -205,49 +147,36 @@ function CreditCard(name) {
 // ----- specific card regexp validation
 // ---------------------------------------
 valid.card = {
-	mastercard: new CreditCard('mastercard'),
-	visa: new CreditCard('visa'),
-	amex: new CreditCard('amex'),
-	maestro: new CreditCard('maestro'),
-	jcb: new CreditCard('jcb'),
-	unionpay: new CreditCard('unionpay'),
-	discover: new CreditCard('discover'),
-	solo: new CreditCard('solo'),
-	carteBlanche: new CreditCard('carteBlanche'),
-	dinersClub: new CreditCard('dinersClub'),
-	lasercard: new CreditCard('lasercard')
+	visa: new CreditCard(/^4[0-9]{12}(?:[0-9]{3})?$/),
+	mastercard: new CreditCard(/^5[1-5][0-9]{14}$/),
+	amex: new CreditCard(/^3[47][0-9]{13}$/),
+	carteBlanche: new CreditCard(/^389[0-9]{11}$|^380[0-9]{11}$/),
+	dinersClub: new CreditCard(/^3(?:0[0-5]|[68][0-9])[0-9]{11}$/),
+	discover: new CreditCard(/^65[4-9][0-9]{13}|64[4-9][0-9]{13}|6011[0-9]{12}|(622(?:12[6-9]|1[3-9][0-9]|[2-8][0-9][0-9]|9[01][0-9]|92[0-5])[0-9]{10})$/),
+	jcb: new CreditCard(/^(?:2131|1800|35\d{3})\d{11}$/),
+	lasercard: new CreditCard(/^(6304|6706|6709|6771)[0-9]{12,15}$/),
+	maestro: new CreditCard(/^(5018|5020|5038|6304|6759|6761|6763|6799)[0-9]{8,19}$/),
+	solo: new CreditCard(/^(6334|6767)[0-9]{12}|(6334|6767)[0-9]{14}|(6334|6767)[0-9]{15}$|^(4903|4905|4911|4936|6333|6759)[0-9]{12}|(4903|4905|4911|4936|6333|6759)[0-9]{14}|(4903|4905|4911|4936|6333|6759)[0-9]{15}|564182[0-9]{10}|564182[0-9]{12}|564182[0-9]{13}|633110[0-9]{10}|633110[0-9]{12}|633110[0-9]{13}$/),
+	unionpay: new CreditCard(/^(62[0-9]{14,17})$/)
 };
 
 
 // ----- generic card validation
-//		--	1. input must be number
+//		--	1. input must be a string
 //		--	2. input length must be valid
-//		--	3. return true if any specific test passes
-//		--	4. perform complex check based on number
+//		--	3. perform check based on number
 // ---------------------------------------
 valid.creditCard = valid.card.generic = function(str) {
 
-	if (!valid._isString(str)) {
-		return false;
-	}
+	// [1] throws here if not a string
+	str = str.toLowerCase();
 
-	// must be valid card length
+	// [2] must be valid card length
 	if (str.length < 13 || str.length > 19) {
 		return false;
 	}
 
-	// should return true if any specific regexp tests pass
-	for (var prop in valid._regexps.creditCard) {
-
-		if (valid._regexps.creditCard.hasOwnProperty(prop)) {
-			if (valid._regexps.creditCard[prop].test(str)) {
-				return true;
-			}
-		}
-
-	}
-
-	// method of testing credit cards from 
+	// [3] method of testing credit cards from 
 	// http://ajax.aspnetcdn.com/ajax/jquery.validate/1.13.1/jquery.validate.js
 	var testNumber = 0;
 
@@ -267,6 +196,3 @@ valid.creditCard = valid.card.generic = function(str) {
 	return (testNumber % 10) === 0;
 
 };
-
-
-module.exports = valid;